calories to float

+++ server/graphql/schema.ts
@@ -11,7 +11,7 @@ const typeDefs = gql`
   }
   type sevenDaysIntake {
     date: String
-    calories: Int
+    calories: Float
   } 
   type User {
     user: String!
